+++ software_esp32/src/Services.cpp
@@ -136,7 +136,7 @@ void CServices::checkGetNtp()
     VdmNet.sntpReachable=VdmNet.checkSntpReachable();
     #ifdef EnvDevelop
         UART_DBG.println("checkGetNtp : "+String(VdmNet.sntpReachable));
-      #endif
+    #endif
     if (VdmNet.sntpReachable) {
       if (VdmSystem.getLocalTime(&VdmNet.startTimeinfo)) {
         if ((VdmNet.startTimeinfo.tm_hour==getNtpHour) && (VdmNet.startTimeinfo.tm_min==getNtpMin)) {
@@ -161,12 +161,14 @@ void CServices::servicesLoop()
 
 void CServices::runOnce() 
 {
-    StmApp.getParametersFromSTM();
+  VdmNet.setupNtp();
+  StmApp.getParametersFromSTM();
 }
 
 void CServices::runOnceDelayed10()
 {
   VdmTask.startApp();
+  GetLastTargetValues();
   VdmNet.startBroker();
   checkGetNtp();
 }
@@ -175,6 +177,18 @@ void CServices::runOnceDelayed60()
 {
   VdmTask.startPIServices();
   VdmTask.piTaskInitiated=true;
+}
+
+void CServices::GetLastTargetValues()
+{
+  for (uint8_t picIdx=0; picIdx<ACTUATOR_COUNT; picIdx++) { 
+    if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[picIdx].controlFlags.active) {
+      if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[picIdx].link==0) { 
+          PiControl[picIdx].valveIndex=picIdx;
+          PiControl[picIdx].reloadPiControl();
+      }
+    }
+  }
 }
 
 void CServices::restartSystem(bool waitQueueFinished) {